PARQUET-455: Fix OS X / Clang compiler warnings
There actually was a legitimate bug fixed here for malformed Parquet files, but we are not yet in a position to write a decent test for it until PARQUET-497. I will make a note on that JIRA.

I also set our Travis CI build to fail on future compiler warnings.

This also closes #15.

Author: Wes McKinney <wes@cloudera.com>

Closes #40 from wesm/PARQUET-455 and squashes the following commits:

a348063 [Wes McKinney] Compiler warnings fail the build
271d71e [Wes McKinney] Fix OS X / Clang compiler warnings