PARQUET-1180: Fix behaviour of num_children element of primitive nodes
Per the parquet.thift spec, for primitive nodes the num_children schema
attibute should remain unset. This is implemeted correctly in parquet-mr
see [1]. However currently parquet-cpp does set the num_children
attribute to 0 if it is a primitive node. This pull requests fixes this
issue and the tests that were relying on this behavior

[1] parquet-mr/parquet-hadoop/src/main/java/org/apache/parquet/format/converter/ParquetMetadataConverter.java

Author: Alec Posney <aposney@atlassian.com>

Closes #427 from Posnet/fix-primite-node-num-children and squashes the following commits:

e5af5c6 [Alec Posney] Fix behavior of num_children element of primitive nodes